requirements.txt
langchain-openai>=0.0.5
langchain-core>=0.1.0
openai>=1.0.0  # Explicit dependency for images API (langchain-openai may not include latest)
tavily-python>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
requests>=2.25.0
httpx>=0.24.0  # For image URL downloads (fallback for DALL-E models)
markdown>=3.4.0
numpy>=1.24.0
jinja2>=3.0.0
beautifulsoup4>=4.12.0
